Use gevent workers rather than sync workers
Users are having problems when trying to upload CSV files depending on
the size of the CSV and their network connection - after a certain
amount of time their request times out. At the moment we use sync
workers, so one of the limited number of request slots on the server is
blocked throughout this step. By switching to gevent workers we can
accept a larger number of connections and increase concurrency, and
therefore also handle long-running uploads more gracefully.