Prevent users from picking invalid db identifiers
The dataset will eventually fail to import if we allow freeform field
names here for the db schema and table name. Let's be a bit restrictive
here (alphanumeric, underscores, dashes, periods) to make sure that the
names are fine. We currently only expect Data Workspace team members to
be uploading data this way, so we can be a bit vague about the
requirements.

We will want to enforce the standards automatically when we release this
to more users, and we have a Trello ticket in the backlog to work on
this.