Fix flaky tests
Some of these tests insert comma-separated values, but the pipeline
factory wasn't told that they were comma-separated so it would randomly
pick either comma-separated or tab-separated. This patch pins them to
comma-separated so that the file is processed correctly. This problem
was found when adding 'invalid header' detection, which fails when
headers include any non-alphanumeric, non-underscore character. If the
pipeline factory was configured as tab-separated, the header would be
something like 'one,two', rather than 'one', 'two'.