Add a route back to processing progress
This patch adds a link back to the 'data processing progress' page from
the pipeline overview page. Previously it wasn't possible to get back to
this page through the UI (though you could hack the URL). Reloading the
'progress' page would also end up spawning a new processor thread,
meaning that the same data file could be processed multiple times. This
moves the spawning step to the validated POST, and leaves the 'GET
progress' view as just returning updates on how the processor is doing.