refactor: Move column_types, delimiter and quote from pipeline to datafile
Columns, their types, delimiter and quote char can vary between different datafiles of the same pipeline. To keep track of what has been used or selected by a user we save them in the DB for each datafile sepeartely instead of once for the pipeline. This allows to restore a previous datafile with a previous configuration.