feature: companies house accounts pipeline
This pipeline allows the processing of XBRL documents submitted by companies
to Companies House. Because the data is only availble as unstructured xml
documents, the pipeline uses XPath expressions to extract the data into a more readable
database table structure. At the end of the processing there is the option to
trigger a dataflow dag to load the data into dataworkspace.