feat: clean up old pipelie processing tasks
A rudimentary implementation of a cleanup job that ensures pipelines
don't get stuck in "processing" step forever. This can happen right now
if the app instance dies while a data file is being processed (either by
data-store-service or data-flow), because the task is simply kicked off
as a background thread on the web worker and there is no retry/recovery
mechanism.

In the future we'll want to move to some more robust background job
handler (e.g. celery), but for now this service is only used within our
team and so that level of robustness is not required.