fix: add prompt to run migrations
Data Store Service is not set up with automatic migrations currently,
because of some historical fears about some tables not being OK to
migrate automatically.

Until we can investigate that, let's at least put a reminder into the
pipeline so that people know it needs to be done manually/with thought.