feat: add pipeline failure page
At the moment, pipeline failures are reported within a green Design
System panel component, which is the same pattern we use for success of
the pipeline. This makes it harder for users to understand the end
result of a pipeline because it's easy to scan the page, see the green
banner, and expect success. We add a separate page for failures states
here to highlight the issue and provide a recommendation - to try
ingesting again.