build(deps): move to Python 3.9 and latest data-engineering-common
To make this and future dependency bumping less painful, this deliberately
unpinned the versions from the requirements.in file. This is a pattern we use
elsewhere, and I think is still fine since all the dependency versions are
still locked in the requirements.txt file which is used during
install/deployment.