feat: older CompaniesHouseAccounts pipeline ingest to _legacy table
There is decreasing confidence in the older v1 Companies House
accounts pipeline within the data-store-service, and so this
will now ingest to a _legacy table in order to use the data for
comparison, but to make the newer incremental pipeline ingest
to the main accounts table. This means that the data ingested
using stream-read-xbrl can be used within the Export Propensity
pipeline.