dev: make it possible again to run data-store-service locally
The previous node version would not install (erroring with messages about
unsupported versions), and was not the same as the one running in CircleCI.
Just changing the existing script to the same node version as in CircleCI also
errored out, so an alternative, and hopefully more robust way is used. This now
seems to work fine locally.

This deliberately does away with the separate script to install node - this
isn't needed and makes it harder to see what's going on when you need to change
things.

This also updates the command to run locally `from docker-compose` to `docker
compose`, which is the current recommended way to run compose file.